HUD and Lives Singleton
- Added a Singleton for the lives.
This makes it so the game keeps track of how many lives you have every time it reloads, so it resets the scene when you die, and eventually, we can make it so it goes to a gameover screen/start screen when you die.
- HUD
Fully implemented the HUD, for now. We might want to move it, though, or resize it, as you go behind it.